+++ pkg/store/cache/memcached.go
@@ -64,7 +64,7 @@ func NewRemoteIndexCache(logger log.Logger, cacheClient cacheutil.RemoteCacheCli
 // The function enqueues the request and returns immediately: the entry will be
 // asynchronously stored in the cache.
 func (c *RemoteIndexCache) StorePostings(blockID ulid.ULID, l labels.Label, v []byte) {
-	key := cacheKey{blockID, cacheKeyPostings(l)}.string()
+	key := cacheKey{blockID.String(), cacheKeyPostings(l)}.string()
 
 	if err := c.memcached.SetAsync(key, v, memcachedDefaultTTL); err != nil {
 		level.Error(c.logger).Log("msg", "failed to cache postings in memcached", "err", err)
@@ -75,16 +75,12 @@ func (c *RemoteIndexCache) StorePostings(blockID ulid.ULID, l labels.Label, v []
 // and returns a map containing cache hits, along with a list of missing keys.
 // In case of error, it logs and return an empty cache hits map.
 func (c *RemoteIndexCache) FetchMultiPostings(ctx context.Context, blockID ulid.ULID, lbls []labels.Label) (hits map[labels.Label][]byte, misses []labels.Label) {
-	// Build the cache keys, while keeping a map between input label and the cache key
-	// so that we can easily reverse it back after the GetMulti().
 	keys := make([]string, 0, len(lbls))
-	keysMapping := map[labels.Label]string{}
 
+	blockIDKey := blockID.String()
 	for _, lbl := range lbls {
-		key := cacheKey{blockID, cacheKeyPostings(lbl)}.string()
-
+		key := cacheKey{blockIDKey, cacheKeyPostings(lbl)}.string()
 		keys = append(keys, key)
-		keysMapping[lbl] = key
 	}
 
 	// Fetch the keys from memcached in a single request.
@@ -96,18 +92,11 @@ func (c *RemoteIndexCache) FetchMultiPostings(ctx context.Context, blockID ulid.
 
 	// Construct the resulting hits map and list of missing keys. We iterate on the input
 	// list of labels to be able to easily create the list of ones in a single iteration.
-	hits = map[labels.Label][]byte{}
-
-	for _, lbl := range lbls {
-		key, ok := keysMapping[lbl]
-		if !ok {
-			level.Error(c.logger).Log("msg", "keys mapping inconsistency found in memcached index cache client", "type", "postings", "label", lbl.Name+":"+lbl.Value)
-			continue
-		}
-
+	hits = make(map[labels.Label][]byte, len(results))
+	for i, lbl := range lbls {
 		// Check if the key has been found in memcached. If not, we add it to the list
 		// of missing keys.
-		value, ok := results[key]
+		value, ok := results[keys[i]]
 		if !ok {
 			misses = append(misses, lbl)
 			continue
@@ -124,7 +113,7 @@ func (c *RemoteIndexCache) FetchMultiPostings(ctx context.Context, blockID ulid.
 // The function enqueues the request and returns immediately: the entry will be
 // asynchronously stored in the cache.
 func (c *RemoteIndexCache) StoreSeries(blockID ulid.ULID, id storage.SeriesRef, v []byte) {
-	key := cacheKey{blockID, cacheKeySeries(id)}.string()
+	key := cacheKey{blockID.String(), cacheKeySeries(id)}.string()
 
 	if err := c.memcached.SetAsync(key, v, memcachedDefaultTTL); err != nil {
 		level.Error(c.logger).Log("msg", "failed to cache series in memcached", "err", err)
@@ -135,16 +124,12 @@ func (c *RemoteIndexCache) StoreSeries(blockID ulid.ULID, id storage.SeriesRef, 
 // and returns a map containing cache hits, along with a list of missing IDs.
 // In case of error, it logs and return an empty cache hits map.
 func (c *RemoteIndexCache) FetchMultiSeries(ctx context.Context, blockID ulid.ULID, ids []storage.SeriesRef) (hits map[storage.SeriesRef][]byte, misses []storage.SeriesRef) {
-	// Build the cache keys, while keeping a map between input id and the cache key
-	// so that we can easily reverse it back after the GetMulti().
 	keys := make([]string, 0, len(ids))
-	keysMapping := map[storage.SeriesRef]string{}
 
+	blockIDKey := blockID.String()
 	for _, id := range ids {
-		key := cacheKey{blockID, cacheKeySeries(id)}.string()
-
+		key := cacheKey{blockIDKey, cacheKeySeries(id)}.string()
 		keys = append(keys, key)
-		keysMapping[id] = key
 	}
 
 	// Fetch the keys from memcached in a single request.
@@ -156,18 +141,11 @@ func (c *RemoteIndexCache) FetchMultiSeries(ctx context.Context, blockID ulid.UL
 
 	// Construct the resulting hits map and list of missing keys. We iterate on the input
 	// list of ids to be able to easily create the list of ones in a single iteration.
-	hits = map[storage.SeriesRef][]byte{}
-
-	for _, id := range ids {
-		key, ok := keysMapping[id]
-		if !ok {
-			level.Error(c.logger).Log("msg", "keys mapping inconsistency found in memcached index cache client", "type", "series", "id", id)
-			continue
-		}
-
+	hits = make(map[storage.SeriesRef][]byte, len(results))
+	for i, id := range ids {
 		// Check if the key has been found in memcached. If not, we add it to the list
 		// of missing keys.
-		value, ok := results[key]
+		value, ok := results[keys[i]]
 		if !ok {
 			misses = append(misses, id)
 			continue